k8s/apps/order-admin/requirements.txt
flask
psycopg2-binary
prometheus_client
gunicorn
